osrepo-legacy/osrepo/pyproject.toml
[tool.poetry]
name = "osrepo"
version = "1.0.0"
description = "an OS image install manager"
authors = ["baris-inandi <68742481+baris-inandi@users.noreply.github.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
inquirer = "^2.9.2"
click = "^8.1.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
